move from decimal to float

+++ Genelife.Domain/Events/Company/PayrollCompleted.cs
@@ -2,4 +2,4 @@ using MassTransit;
 
 namespace Genelife.Domain.Events.Company;
 
-public record PayrollCompleted(Guid CorrelationId, decimal TotalPaid, decimal TaxesPaid) : CorrelatedBy<Guid>;
+public record PayrollCompleted(Guid CorrelationId, float TotalPaid, float TaxesPaid) : CorrelatedBy<Guid>;
