+++ Genelife.Main/Usecases/CalculatePayroll.cs
@@ -6,13 +6,13 @@ namespace Genelife.Main.Usecases;
 
 public class CalculatePayroll
 {
-    public (decimal totalPaid, decimal totalTaxes, List<SalaryPaid> salaryPayments) Execute(
+    public (float totalPaid, float totalTaxes, List<SalaryPaid> salaryPayments) Execute(
         Company company, 
         List<Employee> employees)
     {
         var salaryPayments = new List<SalaryPaid>();
-        decimal totalPaid = 0;
-        decimal totalTaxes = 0;
+        float totalPaid = 0;
+        float totalTaxes = 0;
 
         foreach (var employment in employees.Where(e => e.Status == EmploymentStatus.Active))
         {
